Add func to extend sklearn.datasets.make_* funcs
Try this:

    import xarray_filters.datasets as ds
    import pandas as pd
    help(ds.make_blobs)   # note argument list and docstring
    X, y = ds.make_blobs(astype=pd.DataFrame)

X, y should be dataframes, not arrays. Note that y should be a Series
for things to work correctly with sklearn.

The example above shows where we are going: both docstring and signature
are extensions of the corresponding sklearn.datasets function.

Right now the only extra functionality is the `astype` keyword. We
should polish that a bit, and include other keywords. Like "shape", or
named dimensions, etc. This might get more complicated and require a
refactor, but the current code shows what I'd like to get: make it
clear (through docstring and signature) that we are just extending
sklearn functions.

The challenge is to keep the code readable too. Right now, the resulting
module that is imported is more understandable than the code itself.
Suggestions welcome.