Better support for postprocessing sklearn samplers
A sklearn sampler is one of

    sklearn.datasets.make_blobs
    sklearn.datasets.make_classification
    sklearn.datasets.make_regression

We would like to use those samplers to generate data that will then be
postprocessed/transformed to generate the data structures we want.

To do so, we devised a XyTransformer object that takes the output from
the sklearn functions above as initialization parameters. This object
then provides various to_* functions that allow us to transform the data
generated by sklearn into other data structures (say pandas.DataFrame or
xarray.DataArray objects with certain shapes, labeled coordinates,
dimesions, etc.).

Using this XyTransformer object, we define new functions

    make_blobs = _make_base(sklearn.datasets.make_blobs)
    make_classification = _make_base(sklearn.datasets.make_classification)
    make_regression = _make_base(sklearn.datasets.make_regression)

Where _make_base is a function that takes the output of the sklearn
function passed as argument and creates a XyTransform object. It also
sets an informative docstring for the new make_* functions and sets the
signature of make_* to be the same as its sklearn counterpart. This
facilitates interactive and IDE use.

One can then use, say, make_blobs, to create, for example a new
dataframe of feature names and series of labels with

    simdata = make_blobs()
    simdata.to_dataframe()

The current list of transformation methods implemented in XyTransformer
is

    to_array(self, feature_shape=None)
    to_dataframe(self, feature_names=None, label_name=None)
    to_flat_xarray_with_layers(self, layers=None)