Added astype kw to the make_* funcs
Now we can just call make_blobs(astype) (or any other compatible sklearn
make_* function) to obtain data in the right type.

There are various assertions and checks to minimize the chance that
invalid data is passed to this funciton. We also make it clear that we
cannot convert all make_* functions (see various assetions and the
converted_make_funcs dictionary).