Make datasets.py work with python27
However, the signature of the `make_*` functions exported by datasets.py
is just make_blobs(*args, **kwargs) (using make_blobs as example).

And still, doctests for the datasets.py module fail on python2.7
because

Expected:
    Index(['temp', 'pressure', 'humidity'], dtype='object')
Got:
    Index([u'temp', u'pressure', u'humidity'], dtype='object')

So while there is no error in code, we need to make the doctest robust to those
silly encoding issues.