Make reshape.py pass doctests in py27
Had to put a `doctest: +SKIP` flag on some doctests.

Reason is that the doctest is too useful as documentation, so I'd rather
not butcher it to make it pass. But it fails on py27 because of the
difference in the literal representation of (unicode) strings in Python
2 vs Python 3: u'ball' (py2) vs 'ball' (py3).

Tests _did_ pass on python 3. It would be nice to still run them on
python 3. Maybe we can add our own filter to be run before the tests,
skipping them only under py27.