Fix tests/test_reshape.py
That test was failing under Python 2.7.

All failures were to calls like this (note the strings 'mean' and 'std')

    def excerpt from the function test_aggregations_can_chain():
        ...
        args_kwargs = [['mean', dict(dim='z')], # From 4D to 3D
                       ['std', dict(dim='t')]] # From 3D to 2D
        layers = ['temperature', 'pressure']
        new_X = X.chain(args_kwargs, layers=layers
                       ).to_features(features_layer='temp_pres')

We had an AttributeError in Python2.7 with the test saying that "str type has
no __module__ attribute", because 'mean' and 'std' were being passed ultimately
(via X.chain) to functools.wraps, which calls functools.update_wrapper, that in
turn expected 'mean' to be a __function__ to be wrapped (not a string).

The fix was to turn them into proper functions: 'mean' became MLDataset.mean
and 'std' became MLDataset.std.

    def excerpt from the function test_aggregations_can_chain():
        ...
        args_kwargs = [[MLDataset.mean, dict(dim='z')], # From 4D to 3D
                       [MLDataset.std, dict(dim='t')]] # From 3D to 2D
        layers = ['temperature', 'pressure']
        new_X = X.chain(args_kwargs, layers=layers
                       ).to_features(features_layer='temp_pres')

Interestingly enough, the tests for python3 did not fail. That is because the
code for functools.update_wrapper is different from the python2 correspondent:

In Python 2 (part of the code of functools.update_wrapper, with
wrapped='mean' and attr='__module__'):

    for attr in assigned:
        setattr(wrapper, attr, getattr(wrapped, attr))

In Python3, that became

    for attr in assigned:
        try:
            value = getattr(wrapped, attr)
        except AttributeError:
            pass
        else:
            setattr(wrapper, attr, value)

notice that in the first case, the gettattr(wrapped, attr) call creates an
uncaught AttributeError, while in the py3 version, it's caught and ignored.

__CONCLUSION__: remember to pass real MLData methods (not xarray.DataArray
methods, that doesn't seem to work either) to X.chain.